+++ packages/mobile/src/components/WithPasteAware.tsx
@@ -3,7 +3,7 @@
 import Clipboard from '@react-native-community/clipboard'
 import * as React from 'react'
 import { AppState, NativeEventSubscription, ViewProps } from 'react-native'
-import { deviceIsIos14OrNewer } from 'src/components/utils/IosVersionUtils'
+import { deviceIsIos14OrNewer } from 'src/utils/IosVersionUtils'
 
 interface PasteAwareProps {
   value: string
